pyproject.toml
[tool.poetry]
name = "inpt_glucose"
version = "0.1.0"
description = "Inpatient Glucose Control Assistant"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^1.31.1"
crewai = "^0.16.3"
langchain = "^0.1.8"
openai = "^1.12.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
